fix: reduce react doctor accessibility warnings
