+++ frontend/app/dev/whimsy-tile/page.tsx
@@ -1,3 +1,4 @@
+import type { Metadata } from 'next';
 import type React from 'react';
 import { WHIMSY_PRESETS, whimsyPresetUrl } from '@/lib/whimsy-presets';
 import {
@@ -6,6 +7,11 @@ import {
 	WHIMSY_THEMES,
 	type WhimsyThemeName,
 } from '@/lib/whimsy-tile';
+
+export const metadata: Metadata = {
+	title: 'Whimsy Tile Dev',
+	description: 'Dev-only preview page for generated whimsy tile variants.',
+};
 
 /**
  * Tile dimension in pixels. Same value is used for the SVG `viewBox`, the CSS
