+++ frontend/components/ai-elements/message.tsx
@@ -8,6 +8,7 @@
 
 import type { FileUIPart, UIMessage } from 'ai';
 import { ChevronLeftIcon, ChevronRightIcon, PaperclipIcon, XIcon } from 'lucide-react';
+import Image from 'next/image';
 import type { ComponentProps, HTMLAttributes, ReactElement } from 'react';
 import { createContext, memo, use, useEffect, useState } from 'react';
 import { Streamdown } from 'streamdown';
@@ -340,11 +341,12 @@ export function MessageAttachment({ data, className, onRemove, ...props }: Messa
 		>
 			{isImage ? (
 				<>
-					<img
+					<Image
 						alt={filename || 'attachment'}
 						className="size-full object-cover"
 						height={100}
 						src={data.url}
+						unoptimized
 						width={100}
 					/>
 					{onRemove && (
