+++ frontend/components/ai-elements/prompt-input-attachments.tsx
@@ -9,6 +9,7 @@
 import { DropdownMenuItem } from '@octavian-tocan/react-dropdown';
 import type { FileUIPart } from 'ai';
 import { ImageIcon, PaperclipIcon, XIcon } from 'lucide-react';
+import Image from 'next/image';
 import { type ComponentProps, Fragment, type HTMLAttributes, type ReactNode } from 'react';
 import { Button } from '@/components/ui/button';
 import { cn } from '@/lib/utils';
@@ -47,11 +48,12 @@ export function PromptInputAttachment({ data, className, ...props }: PromptInput
 					<div className="relative size-5 shrink-0">
 						<div className="absolute inset-0 flex size-5 items-center justify-center overflow-hidden rounded bg-background transition-opacity group-hover:opacity-0">
 							{isImage ? (
-								<img
+								<Image
 									alt={filename || 'attachment'}
 									className="size-5 object-cover"
 									height={20}
 									src={data.url}
+									unoptimized
 									width={20}
 								/>
 							) : (
@@ -82,11 +84,12 @@ export function PromptInputAttachment({ data, className, ...props }: PromptInput
 				<div className="w-auto space-y-3">
 					{isImage && (
 						<div className="flex max-size-96 items-center justify-center overflow-hidden rounded-md border">
-							<img
+							<Image
 								alt={filename || 'attachment preview'}
 								className="max-h-full max-w-full object-contain"
 								height={384}
 								src={data.url}
+								unoptimized
 								width={448}
 							/>
 						</div>
