+++ frontend/components/signup-form.tsx
@@ -6,6 +6,7 @@
 
 'use client';
 
+import Link from 'next/link';
 import { useRouter } from 'next/navigation';
 import { useId, useState } from 'react';
 import { Button } from '@/components/ui/button';
@@ -165,7 +166,7 @@ export function SignupForm({ ...props }: React.ComponentProps<typeof Card>) {
                                     Sign up with Google
                                 </Button> */}
 								<FieldDescription className="px-6 text-center">
-									Already have an account? <a href="/login">Sign in</a>
+									Already have an account? <Link href="/login">Sign in</Link>
 								</FieldDescription>
 							</Field>
 						</FieldGroup>
