+++ frontend/features/onboarding/v2/step-personality.tsx
@@ -41,7 +41,7 @@ export function StepPersonality({
 					size="lg"
 					type="button"
 				>
-					Continue
+					Save personality
 				</Button>
 			}
 			subtitle="Pick a personality."
