+++ frontend/features/projects/components/ProjectsList.tsx
@@ -276,7 +276,6 @@ function RenameProjectModalInner({
 				</ModalDescription>
 				<AppFormRow htmlFor={inputId} label="Project name" labelVisibility="sr-only">
 					<Input
-						autoFocus
 						id={inputId}
 						maxLength={255}
 						onChange={(event) => setDraft(event.target.value)}
